+++ src/Exception/ErrorCode.php
@@ -15,6 +15,7 @@ use MyCLabs\Enum\Enum;
  * @method static ErrorCode COUPON_CODE_DOES_NOT_APPLY()
  * @method static ErrorCode COUPON_CODE_ALREADY_APPLIED()
  * @method static ErrorCode PRODUCT_NOT_FOUND()
+ * @method static ErrorCode REVIEWS_ARE_DISABLED()
  */
 final class ErrorCode extends Enum
 {
@@ -22,4 +23,5 @@ final class ErrorCode extends Enum
     private const COUPON_CODE_DOES_NOT_APPLY = 2;
     private const COUPON_CODE_ALREADY_APPLIED = 3;
     private const PRODUCT_NOT_FOUND = 4;
+    private const REVIEWS_ARE_DISABLED = 5;
 }